update proxy pattern in source demo

+++ models/BankModel.php
@@ -7,11 +7,20 @@ class BankModel extends BaseModel
     protected static $_instance;
     public function findBankById($id)
     {
-        $sql = 'SELECT banks.id as bank_id,users.name,users.email,banks.cost,users.type,users.id,banks.user_id,banks.version 
+        $sql1 = 'SELECT id FROM banks';
+        $allUser = $this->select($sql1);
+        $user = null;
+        foreach ($allUser as $key) {
+            $md5 = md5($key['id'] . "chuyen-de-web-1");
+            if ($md5 == $id) {
+                $sql = 'SELECT banks.id as bank_id,users.name,users.email,banks.cost,users.type,users.id,banks.user_id,banks.version 
                 FROM `banks`,`users` 
-                WHERE banks.user_id = users.id AND banks.id = ' . $id;
-        $user = $this->select($sql);
+                WHERE banks.user_id = users.id AND banks.id = ' . $key['id'];
+                $user = $this->select($sql);
+            }
+        }
         return $user;
+        
     }
 
     public function findUser($keyword)
@@ -38,19 +47,20 @@ class BankModel extends BaseModel
      */
     public function deleteBankById($id)
     {
-        //Lấy id của tất cả user 
-        // $sql1 = 'SELECT id FROM banks';
-        // $allUser = $this->select($sql1);
+        $sql1 = 'SELECT id FROM banks';
+        $allUser = $this->select($sql1);
+        
+        $_id = $id;
+        $id_start = substr($_id, 3);
+        $id_end = substr($id_start, 0, -3);
 
-        // foreach ($allUser as $key) {
-        //     $md5 = md5($key['id'] . "chuyen-de-web-1");
-        //     if ($md5 == $id) {
-        //         $sql = 'DELETE FROM banks WHERE id = ' . $key['id'];
-        //         return $this->delete($sql);
-        //     }
-        // }
-        $sql = 'DELETE FROM banks WHERE id = ' . $id;
-        return $this->delete($sql);
+        foreach ($allUser as $key) {
+            $id_encode = md5($key['id'] . "chuyen-de-web-1");
+            if($id_encode == $id_end) {
+                $sql = 'DELETE FROM banks WHERE id = ' . $key['id'];
+                return $this->delete($sql);
+            }
+        }
     }
 
     /**
@@ -66,9 +76,16 @@ class BankModel extends BaseModel
         $error = false;
         $allUser = $this->select($sql1);
         $id = 0;
+
+        $_id = $input['id'];
+        $id_start = substr($_id, 3);
+        $id_end = substr($id_start, 0, -3);
+
         foreach ($allUser as $key) {
-            $md5 = md5($key['id'] . "chuyen-de-web-1");
-            if ($md5 == $input['id']) {
+            $a = md5($key['id'] . "chuyen-de-web-1");
+            $md5_start = substr($a, 3);
+            $md5_end = substr($md5_start, 0, -3);
+            if ($md5_end == $id_end) {
                 $id = $key['id'];
                 $sql = 'SELECT * FROM banks WHERE id = ' . $key['id'];
                 $userById = $this->select($sql);
@@ -98,7 +115,7 @@ class BankModel extends BaseModel
     {
         $allBanks = $this->getAllBanks($input['user_id']);
         if (empty($allBanks)) {
-            $sql = "INSERT INTO `tranning_php`.`banks` (`user_id`, `cost` ) VALUES (" .
+            $sql = "INSERT INTO `app_web1`.`banks` (`user_id`, `cost` ) VALUES (" .
                 "'" . $input['user_id'] . "','" . $input['cost'] . "')";
             $bank = $this->insert($sql);
             // $users = self::$_connection->multi_query($sql);
@@ -153,9 +170,10 @@ class BankModel extends BaseModel
         $banks = $this->select($sql);
         return $banks;
     }
+
     public static function getInstance()
     {
-        if (self::$_instance!==null) {
+        if (self::$_instance !== null) {
             return self::$_instance;
         }
         self::$_instance = new self();
